+++ packages/learn-card-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # learn-card-core
 
+## 9.4.30
+
+### Patch Changes
+
+-   Updated dependencies [[`80bce3e745d8e3dcca0e1c07337d97ec13c4b417`](https://github.com/learningeconomy/LearnCard/commit/80bce3e745d8e3dcca0e1c07337d97ec13c4b417)]:
+    -   @learncard/helpers@1.3.12
+
 ## 9.4.29
 
 ### Patch Changes